pyproject.toml
[project]
name = "fastgen"
version = "0.1.0"
description = "Scaffold your projects effortlessly with FastGen"
authors = [
    {name = "Ali Zamini"}
]
packages = [{include = "fastgen"}, {include = "core"}, {include = "utils"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "rich (>=13.9.4,<14.0.0)"
]

[tool.poetry.scripts]
fastgen = "fastgen.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
